+++ security-ldap/src/main/java/io/micronaut/security/ldap/configuration/LdapConfiguration.java
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.configuration.security.ldap.configuration;
+package io.micronaut.security.ldap.configuration;
 
-import io.micronaut.configuration.security.ldap.context.ContextConfigurationContextSettings;
-import io.micronaut.configuration.security.ldap.context.ContextSettings;
-import io.micronaut.configuration.security.ldap.context.SearchSettings;
+import io.micronaut.security.ldap.context.ContextConfigurationContextSettings;
+import io.micronaut.security.ldap.context.ContextSettings;
+import io.micronaut.security.ldap.context.SearchSettings;
 import io.micronaut.context.annotation.ConfigurationProperties;
 import io.micronaut.context.annotation.EachProperty;
 import io.micronaut.context.annotation.Parameter;
