+++ security-ldap/src/main/java/io/micronaut/security/ldap/context/ContextConfigurationContextSettings.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.configuration.security.ldap.context;
+package io.micronaut.security.ldap.context;
 
-import io.micronaut.configuration.security.ldap.configuration.LdapConfiguration;
+import io.micronaut.security.ldap.configuration.LdapConfiguration;
 
 import java.util.Map;
