+++ security-ldap/src/main/java/io/micronaut/security/ldap/context/ContextSettings.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.configuration.security.ldap.context;
+package io.micronaut.security.ldap.context;
 
 import java.util.Collections;
 import java.util.Map;
