+++ security-ldap/src/test/groovy/io/micronaut/security/ldap/DisableLdapSpec.groovy
@@ -1,6 +1,6 @@
 package io.micronaut.security.ldap
 
-import io.micronaut.configuration.security.ldap.LdapAuthenticationProvider
+
 import io.micronaut.context.ApplicationContext
 import io.micronaut.context.exceptions.NoSuchBeanException
 import io.micronaut.inject.qualifiers.Qualifiers
